Remove 'debian | ubuntu | check if oracle license is already accepted'
The task is unnecessary; it only saves us from running 'debian | set license as accepted'.
However, that task is already idempotent; if the license is already accepted, it does nothing.

Furthermore, the deleted task was problematic because it used 'debconf-get-selections', which
might not be installed (it wasn't for me in Ubuntu 16.04). Rather than installing 'debconf-utils',
why not just nuke the useless task?